Refactor workout json loading. Fix file name matching bug
