+++ apps/buffgym/buffgym-workout.js
@@ -27,8 +27,30 @@ exports = class Workout {
     return !!this.completed;
   }
 
-  static fromJSON(workout) {
+  static fromJSON(workoutJSON) {
+    const Set = require("buffgym-set.js");
+    const Exercise = require("buffgym-exercise.js");
+    const workout = new this({
+      title: workoutJSON.title,
+    });
+    const exercises = workoutJSON.exercises.map(exerciseJSON => {
+      const exercise = new Exercise({
+        title: exerciseJSON.title,
+        weight: exerciseJSON.weight,
+        weightIncrement: exerciseJSON.weightIncrement,
+        unit: exerciseJSON.unit,
+        restPeriod: exerciseJSON.restPeriod,
+      });
+      exerciseJSON.sets.forEach(setJSON => {
+        exercise.addSet(new Set(setJSON));
+      });
 
+      return exercise;
+    });
+
+    workout.addExercises(exercises);
+
+    return workout;
   }
 
   toJSON() {
@@ -38,6 +60,7 @@ exports = class Workout {
         return {
           title: exercise.title,
           weight: exercise.weight,
+          weightIncrement: exercise.weightIncrement,
           unit: exercise.unit,
           sets: exercise.sets.map(set => set.maxReps),
           restPeriod: exercise.restPeriod,
